Split graph_calcs and nw_calcs
